+++ core/src/main/java/com/bestproject/main/ObjectFragment/ParryHitbox.java
@@ -2,7 +2,7 @@ package com.bestproject.main.ObjectFragment;
 
 public class ParryHitbox extends ATKHITBOX {
     boolean parry=true;
-    public ParryHitbox(double x, double y, double z, double width, double thickness, double height, float invincib_frame, float[] damage, boolean isEnemyTarget) {
+    public ParryHitbox(float x, float y, float z, float width, float thickness, float height, float invincib_frame, float[] damage, boolean isEnemyTarget) {
         super(x, y, z, width, thickness, height, invincib_frame, damage, isEnemyTarget);
     }
     @Override
