+++ core/src/main/java/com/bestproject/main/ObjectFragment/SlashHitbox.java
@@ -1,8 +1,8 @@
 package com.bestproject.main.ObjectFragment;
 
 public class SlashHitbox extends ATKHITBOX{
-    public SlashHitbox(double x, double y, double z, double width, double thickness, double height, float invincib_frame, float[] damage, boolean isEnemyTarget) {
+    public SlashHitbox(float x, float y, float z, float width, float thickness, float height, float invincib_frame, float[] damage, boolean isEnemyTarget) {
         super(x, y, z, width, thickness, height, invincib_frame, damage, isEnemyTarget);
     }
-    
+
 }
